Tasks: updating simplified; no more templating.
To make it simpler to update the tasks, they will now use values from `pyproject.toml` if needed instead of from `.cookiecutter.json`. This removes both the need for updating two files when using the tasks and for the more complicated updating step involving templating (which also removes the need for `cookiecutter` as a dev-dependency). Additionally, since the tasks file has no project-specific information embedded in it, it is more portable: you can just copy the file into an existing project if you want to use it (provided you have a `pyproject.toml` with any needed information; so far this is just the project name).