Hooks: `git init`, `poetry install`, `pre-commit`.
In the post-gen hook, we initialize a `git` repo, have `poetry` install dependences and create a virtual environment, and install the `pre-commit` hook. There's a new `cookiecutter` option to skip the `poetry` steps (the install one is slow), which we do for the demo project.

The README has also been updated with some information about the formatting included in this template.