Tasks: add `clean`, use custom invoke fork.
The `clean` task needs to pass arguments through to `pytest`, which `invoke` doesn't support well. My fork has both the fix that was already required to make things work on Python 3 as well as a few changes that allow tasks to work with unknown arguments for when you can't explicitly specify all of them.